requirements.txt
streamlit
pdfplumber
pytesseract
Pillow
langchain_community
langchain
numpy
pypdf
sentence-transformers
transformers
langchain
langchain-groq
unstructured
typing-extensions==4.13.0
rapidocr-onnxruntime
qdrant-client
langchain-qdrant
langchain-huggingface
